Instructor DELETE DONE

+++ RazorPages/RazorPages/Pages/Instructors/Delete.cshtml.cs
@@ -49,13 +49,25 @@ namespace RazorPages.Pages.Instructors
                 return NotFound();
             }
 
-            var instructor = await _context.Instructors.FindAsync(id);
+			Instructor instructor = await _context.Instructors
+											.Include(i => i.Courses)
+											.SingleAsync(i => i.ID == id);
+			if(instructor == null)return RedirectToPage("./Index");
+
+			List<Department> departments = await _context.Departments
+												.Where(d => d.InstructorID==id)
+												.ToListAsync();
+			departments.ForEach(d => d.InstructorID = null);
+			_context.Instructors.Remove(instructor);
+			await _context.SaveChangesAsync();
+
+            /*var instructor = await _context.Instructors.FindAsync(id);
             if (instructor != null)
             {
                 Instructor = instructor;
                 _context.Instructors.Remove(Instructor);
                 await _context.SaveChangesAsync();
-            }
+            }*/
 
             return RedirectToPage("./Index");
         }
